KnowledgeExpansion/requirements.txt
pytorch
pytorch-cuda
torchvision
numpy
matplotlib
tensorboard
tensorboardX
tqdm
scipy
monai
ipykernel
json
glob